physical_ai_textbook/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
google-generativeai
qdrant-client==1.7.3
psycopg2-binary==2.9.9
httpx==0.26.0
sqlalchemy==2.0.25
tiktoken==0.5.2
